requirements.txt
fastapi
uvicorn
pydantic
passlib
python-jose
requests
pytest
pytest-cov